refactor(db): consolidate initial migration into single 20260429_01 revision
Collapse the two pre-history migrations (fbeeb04c3d36 +
61913ced1086) into a single 20260429_01 revision now that the table
has no production deployments and no rows to backfill. Without
existing data the create-default-then-drop-default workaround for
NOT NULL columns is unnecessary, so the upgrade reads as a single
op.create_table mirroring db.py's metadata.

Switch to date-prefixed sequential revision ids (YYYYMMDD_NN) for
human-readable filenames and natural lexical ordering.

Drive-by: align tenant_id and model column lengths with the
SqlAlchemy table definition (String(255)); the original v1
migration declared them unbounded.
